move nonEmptyArray, etc to common
