+++ packages/common/src/functions.test.ts
@@ -8,8 +8,10 @@ import {
 	parseSecretJson,
 	partition,
 	stringToSeverity,
+	toNonEmptyArray,
 	topicMonitoringProductionTagCtas,
 } from './functions';
+import type { NonEmptyArray } from './types';
 
 function isValidUrl(str: string) {
 	try {
@@ -199,5 +201,18 @@ describe('stringToSeverity', () => {
 		expect(stringToSeverity('medium')).toBe('medium');
 		expect(stringToSeverity('high')).toBe('high');
 		expect(stringToSeverity('critical')).toBe('critical');
+	});
+});
+
+describe('Failure on empty arrays', () => {
+	test('throw an error if input is an empty array', () => {
+		const emptyArray: string[] = [];
+		const nonEmptyArray: string[] = ['a', 'b'];
+		const typedNonEmptyArray: NonEmptyArray<string> = ['a', 'b'];
+
+		expect(() => toNonEmptyArray(emptyArray)).toThrow();
+		expect(() => toNonEmptyArray(nonEmptyArray)).not.toThrow();
+		expect(toNonEmptyArray(nonEmptyArray)).toEqual(nonEmptyArray);
+		expect(toNonEmptyArray(nonEmptyArray)).toEqual(typedNonEmptyArray);
 	});
 });
