+++ packages/repocop/src/utils.ts
@@ -1,15 +1,8 @@
 import type { RepocopVulnerability } from 'common/src/types';
-import type { NonEmptyArray, Repository } from './types';
+import type { Repository } from './types';
 
 export function isProduction(repo: Repository) {
 	return repo.topics.includes('production') && !repo.archived;
-}
-
-export function toNonEmptyArray<T>(value: T[]): NonEmptyArray<T> {
-	if (value.length === 0) {
-		throw new Error(`Expected a non-empty array. Source table may be empty.`);
-	}
-	return value as NonEmptyArray<T>;
 }
 
 const criticalFirstPredicate = (x: RepocopVulnerability) =>
